Refactored MapArea and StaticShadow to inherit from Entity.
By this, they can also be added to the environment by a MapObjectLoader and it is easier to manage them in the editor.
Entities with rendertype NONE are now also contained by the environment.

Closes #33

+++ src/de/gurkenlabs/litiengine/environment/tilemap/StaticShadow.java
@@ -19,7 +19,7 @@ public class StaticShadow extends MapArea {
 
   private StaticShadowType shadowType;
 
-  public StaticShadow(int id, String name, double x, double y, double width, double height, StaticShadowType shadowType) {
+  public StaticShadow(int id, String name, double x, double y, float width, float height, StaticShadowType shadowType) {
     super(id, name, x, y, width, height);
     this.setShadowType(shadowType);
   }
